Update Product Page list

+++ Services/order/views.py
@@ -3,9 +3,10 @@ from rest_framework.response import Response
 from django.db.models import Q
 from rest_framework import status
 from rest_framework.exceptions import ValidationError
+import json
 
 from .models import Cart,Favorite,Order,CartProduct
-from products.models import Product
+from products.models import Product,ProductVariant
 from .serializers import CartSerializer,FavoriteSerializer,OrderSerializer
 
 class CartView(APIView):
@@ -23,53 +24,66 @@ class AddUpdateCartView(CartView):
 
     def post(self,request):
         user=request.user
+        payload = request.data
+        for product_item in payload:
+            product_id = product_item.get('product_id')
+            quantity = product_item.get('quantity', 1)
+            product_variant_id = product_item.get('product_variant_id')
+            size = product_item.get('size')
 
-        product_id = request.data.get('product_id')
-        quantity = request.data.get('quantity', 1)
-
-        if not product_id:
-            return Response({"message": "product_id is required"}, status=status.HTTP_400_BAD_REQUEST)
-        
-        if quantity is None or quantity < 0:
-            return Response({"message": "Invalid quantity"}, status=status.HTTP_400_BAD_REQUEST)
-        
-        if user.is_authenticated:
-            try:
-                product = Product.objects.get(uid=product_id)
-            except Product.DoesNotExist:
-                return Response({"message": "Invalid product"}, status=status.HTTP_400_BAD_REQUEST)
+            if not product_id:
+                return Response({"message": "product_id is required"}, status=status.HTTP_400_BAD_REQUEST)
             
-        
-            # Check if the product is already in the user's cart
-            # Separate Cart and created objects from the tuple
-            cart, created = Cart.objects.get_or_create(user=user)
+            if not product_variant_id:
+                return Response({"message": "product_variant_id is required"}, status=status.HTTP_400_BAD_REQUEST)
             
-
-            if quantity == 0:
-                # If quantity is 0, delete the product from the cart if it exists
-                cart_product = CartProduct.objects.filter(cart=cart, product=product).first()
-                if cart_product:
-                    #delete cart product
-                    cart_product.delete()
-                    return Response("Product deleted from the cart.", status=status.HTTP_204_NO_CONTENT)
-                else:
-                    return Response("Product not found in the cart.", status=status.HTTP_404_NOT_FOUND)
-
-            #Create Cart PRoduct
-            cart_product, created = CartProduct.objects.get_or_create(cart=cart, product=product)
-            if not created:
-                cart_product.quantity += quantity
-                cart_product.save()
-                return Response("Quantity updated in the cart.", status=status.HTTP_200_OK)
+            if not size:
+                return Response({"message": "size is required"}, status=status.HTTP_400_BAD_REQUEST)
             
+            if quantity is None or quantity < 0:
+                return Response({"message": "Invalid quantity"}, status=status.HTTP_400_BAD_REQUEST)
+            
+            if user.is_authenticated:
+                try:
+                    product = Product.objects.get(uid=product_id)
+                except Product.DoesNotExist:
+                    return Response({"message": "Invalid product"}, status=status.HTTP_400_BAD_REQUEST)
+                
+                try:
+                    product_variant = ProductVariant.objects.get(uid=product_variant_id)
+                except ProductVariant.DoesNotExist:
+                    return Response({"message": "Invalid product variant"}, status=status.HTTP_400_BAD_REQUEST)
+                
+            
+                # Check if the product is already in the user's cart
+                # Separate Cart and created objects from the tuple
+                cart, created = Cart.objects.get_or_create(user=user)
+                
+
+                if quantity == 0:
+                    # If quantity is 0, delete the product from the cart if it exists
+                    cart_product = CartProduct.objects.filter(cart=cart, product=product, product_variant = product_variant).first()
+                    if cart_product:
+                        #delete cart product
+                        cart_product.delete()
+                        # return Response("Product deleted from the cart.", status=status.HTTP_204_NO_CONTENT)
+                    else:
+                        return Response("Product not found in the cart.", status=status.HTTP_404_NOT_FOUND)
+
+                #Create Cart PRoduct
+                cart_product, created = CartProduct.objects.update_or_create(cart=cart, product=product,
+                                                                          product_variant = product_variant,
+                                                                          size = size,defaults={"quantity": quantity})
+                # if not created:
+                #     cart_product.quantity = quantity
+                #     cart_product.save()
+                #     return Response("Quantity updated in the cart.", status=status.HTTP_200_OK)
+                
             else:
-                # If the product is not in the cart, set the quantity and return the addition response
-                cart_product.quantity = quantity
-                cart_product.save()
-                return Response("Product added to the cart.", status=status.HTTP_201_CREATED)
+                return Response({"msg":"User not Logged In","status":"failed"},status=status.HTTP_401_UNAUTHORIZED)
 
-        else:
-            return Response({"msg":"User not Logged In","status":"failed"},status=status.HTTP_401_UNAUTHORIZED)
+        return Response("Product added to the cart.", status=status.HTTP_201_CREATED)
+        
 
 class FavoriteView(APIView):
     #need to update it in future as User model is created
